Create install-network-printers
This PowerShell script makes a web-request call to the URL of a network printer and searches for the device name via a regular expression for the HTML element, then parses the data and with the reference workbook, installs the printers and prompts the user to uninstall if necessary.  The workbook's ReferencePool sheet is used to and content to the end of the printer name, and the DriveMappings sheet is used to match the regular expression output to DriveName and DriverPath needed for installing the printer.